+++ src/components/components.ts
@@ -1,4 +1,4 @@
-/* Web components. File generated at Sun, 23 Aug 2026 22:57:20 GMT. */
+/* Web components. File generated at Sat, 29 Aug 2026 01:34:16 GMT. */
 export * from './batch-code-editor/batch-code-editor.component';
 export * from './batch-output/batch-output.component';
 export * from './batch-shell/batch-shell.component';